dom/base/test/test_getTranslationNodes_limit.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for nsIDOMWindowUtils.getTranslationNodes</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="runTest()">
<script type="application/javascript">
  var utils = SpecialPowers.wrap(window).
              QueryInterface(SpecialPowers.Ci.nsIInterfaceRequestor).
              getInterface(SpecialPowers.Ci.nsIDOMWindowUtils);

  function runTest() {
    isnot(utils, null, "nsIDOMWindowUtils");

    for (var i = 0; i < 16000; i++) {
      var text = document.createTextNode("a");
      var node = document.createElement("b");
      node.appendChild(text);
      document.body.appendChild(node);
    }

    var translationRoots = utils.getTranslationNodes(document.body);
    is (translationRoots.length, 15000, "Translation nodes were limited to 15000 nodes.");

    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
</script>
</body>
</html>
